add paths
