--- dags/callables/log_model.py
@@ -1,17 +1,0 @@
-import mlflow
-import mlflow.pytorch  # or mlflow.tensorflow / mlflow.sklearn depending on your model
-
-def log_model():
-    print("Logging model to MLflow... (mock)")
-    with mlflow.start_run(run_name="Concrete_Crack_Classifier"):
-        model = torch.load('/path/to/your/latest_model.pth', map_location='cpu')  # Adjust path
-        mlflow.pytorch.log_model(
-            pytorch_model=model,
-            artifact_path="model",
-            registered_model_name="ConcreteCrackClassifierModel"
-        )
-        mlflow.log_metric("accuracy", 0.92)
-        mlflow.log_param("optimizer", "Adam")
-        mlflow.log_param("learning_rate", 0.001)
-        mlflow.log_param("epochs", 20)
-    Variable.set("last_retrain_time", datetime.utcnow().isoformat())
